// apps/web/src/gen/types/RecebimentosDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type RecebimentosDto = {
  /**
   * @type number
   */
  mdr: number
  /**
   * @type number
   */
  tpv: number
  /**
   * @type number
   */
  rav: number
  /**
   * @type number
   */
  liquido: number
  /**
   * @type number
   */
  estabelecimento_id: number
  /**
   * @type number
   */
  taxa_comissao: number
  /**
   * @type string
   */
  numero_documento: string
  /**
   * @type string
   */
  razao_social: string
  /**
   * @type string
   */
  nome: string
  /**
   * @type string
   */
  vendedor_id: string
  /**
   * @type number
   */
  status: number
  /**
   * @type string, date-time
   */
  atribuido_em: string
}